% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reikn_einingar.R
\name{reikn_einingar}
\alias{reikn_einingar}
\title{Reiknar sjópeningaeiningar útfrá forsendum í töflu.}
\usage{
reikn_einingar(tafla)
}
\arguments{
\item{tafla}{Sjópeningatafla, úttak fallsins 'sjopengR::sjopeng'.}
}
\value{
Yfirlit sjópeningaeininga skipt á launaliði.
}
\description{
Reiknar sjópeningaeiningar útfrá forsendum í töflu, í einingum
yfirvinnutíma skipt á launaliðina sjópeninga, óþægindaálag I og óþægindalág II.
}
\references{
https://www.fin.is/media/samningar/Kjarasamningur-FIN-og-rikis-28.februar-2018.pdf,
https://www.fin.is/media/samningar/Hafro-2017.pdf
}
