% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reikn_sjpng.R
\name{reikn_sjpng}
\alias{reikn_sjpng}
\title{Reiknar sjópeningar útfrá forsendum í töflu.}
\usage{
reikn_sjpng(tafla, manadar_laun = 1)
}
\arguments{
\item{tafla}{Sjópeningatafla, úttak fallsins 'sjopengR::sjopeng'.}

\item{manadar_laun}{Sjálfgefið 1, annar hægt að gefa upp kaupið sitt.}
}
\value{
Yfirlit sjópeninga skipt á launaliði.
}
\description{
Reiknar sjópeningar útfrá forsendum í töflu, í einingum mánaðarlauna,
eða í krónum ef grunnlaun eru gefin
með því að nota fallið 'holiday' í pakkanum 'timeDate'.
ATH að stórhátíðadögum í upptalningu kjarasamnings er sleppt.
}
\references{
https://www.fin.is/media/samningar/Kjarasamningur-FIN-og-rikis-28.februar-2018.pdf
}
