% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serstakir.R
\name{serstakir}
\alias{serstakir}
\title{Finnur sérstak frídaga gefins árs.}
\usage{
serstakir(yyyy)
}
\arguments{
\item{yyyy}{Ár.}
}
\value{
Stórhátíðardagar gefið ár.
}
\description{
Finnur sérstaka frídaga skv kjarasamningi fyrir gefið ár
með því að nota fallið 'holiday' í pakkanum 'timeDate'.
ATH að stórhátíðadögum í upptalningu kjarasamnings er sleppt.
}
\references{
https://www.fin.is/media/samningar/Kjarasamningur-FIN-og-rikis-28.februar-2018.pdf
}
