% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjopeng.R
\name{sjopeng}
\alias{sjopeng}
\title{Reiknar sjópeninga útfrá spönn daga og fleiri þáttum.}
\usage{
sjopeng(upphaf, endir, fjarvist_hefst = 3, leidangursstjorn = 0, leiguskip = 0)
}
\arguments{
\item{upphaf, endir}{Upphafs og lokadagur leiðangurs sem ISO-dagsetningar strengur (YYYY-MM-DD).}

\item{fjarvist_hefst}{Dagur í leiðangri sem fjarvistaruppbót byrjar, þriðji dagur ef ekkert hefur verið róið}

\item{leidangursstjorn}{Leiðangurstjóraálag, sjálgefið 0, 0.5 fyrir 1-2 leiðangurmenn, 1 fyrir 4 og fleiri.}

\item{leiguskip}{Er róið á leiguskipi og greitt álag? Sjálgefið 0, en 1/2 tími ef um það er að ræða.}
}
\value{
Nokkuð stór tafla með forsendum sjópengingareiknings.
}
\description{
Reiknar sjópeninga útfrá spönn daga, helgum og helgidögum skv sjókjarasamning Hafró.
Stuðst er við fallið 'holiday' í pakkanum 'timeDate'.
}
\details{
Yfirvinna er greidd með tímakaupi. Tímakaup fyrir yfirvinnu í hverjum launaflokki
er 1,0385\% af mánaðarlaunum starfsmann.

Öll vinna sem unnin er á stórhátíðum skv. gr. 2.1.4.3 greiðist með tímakaupi
sem nemur 1,375\% af mánaðarlaunum starfsmanns.

Sjópeningar - virkir dagar: 4,92 fyrstu tvo dagana og síðan 5,92 með fjarvistaruppbót.

Sjópeningar - helgar: 12,92 fyrstu tvo dagana og síðan 13,92 með fjarvistaruppbót.

Fjarvistaruppbót er 1 klst. pr. dag frá og með 3. degi samfelldrar fjarvistar,
eða frá og með 6. fjarvistardegi í hverjum almanaksmánuði.

Óþægindaálag I (33.33\%) - 3 klst. á dag mánudaga til fimmtudaga.

Óþægindaálag II (55\%) - 3 klst. á dag á föstudögum.

Yfirvinna fyrir leiðangursstjóra í ferðum með þremur eða fleiri er 1 klst á dag
(stórhát.kaup ef rauður dagur).
Yfirvinna fyrir leiðangursstjóra í ferðum með færri en þremur er 0,5 klst á dag
(stórhát.kaup ef rauður dagur).

Ennfremur er litið svo á að allt sem greitt er með yfirvinnutímum
sé á stórhátíðarálgi ef þannig ber undir (óþægindaálag og fjarvistaruppbót auk Ó-álaganna.

Yfirvinna fyrir að vera á öðrum skipum en Hafró er 0,5 klst. á dag

Ef sérstakur frídagur fellur á mánudag til föstudags eru greiddir sjópeningar
eins og það væri helgi (12,92 eða 13,92 - Ath. ekkert óþ.álag.).

Ef stórhátíðardagur fellur á helgi skal greiða sjópeningana (12,92 eða 13,92)
með stórhátíðarkaupi.
}
\references{
https://www.fin.is/media/samningar/Kjarasamningur-FIN-og-rikis-28.februar-2018.pdf,
https://www.fin.is/media/samningar/Hafro-2017.pdf
}
