% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjopengR-package.R
\docType{package}
\name{sjopengR-package}
\alias{sjopengR}
\alias{sjopengR-package}
\title{sjopengR: Reiknar sjópeningar skv sjókjarasamning Hafró og miðlægum kjarasamningi}
\description{
Reiknar sjópeningar skv sjókjarasamning Hafró og miðlægum kjarasamningi, með óþæginda-, fjarvista- og leiðangursstjóraálögum, stórhátíðakaupi og helgartaxta á sérstökum frídögum sem fundnir eru um með 'timeDate::holiday'.
}
\author{
\strong{Maintainer}: Sigurður Þór Jónsson \email{sigurdur.thor.jonsson@gmail.com} (\href{https://orcid.org/0000-0001-5550-0326}{ORCID})

Authors:
\itemize{
  \item Arnþór Bragi Kristinsson
}

}
\keyword{internal}
