% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storhatidir.R
\name{storhatidir}
\alias{storhatidir}
\title{Finnur stórhátíðardaga gefins árs.}
\usage{
storhatidir(yyyy)
}
\arguments{
\item{yyyy}{Ár.}
}
\value{
Stórhátíðardagar gefið ár.
}
\description{
Finnur stórhátíðardaga skv kjarasamningi fyrir gefið ár
}
\references{
https://www.fin.is/media/samningar/Kjarasamningur-FIN-og-rikis-28.februar-2018.pdf
}
